feat: PortfolioCalculator에 평균 매입가 계산 메서드 구현
