+++ src/main/java/co/fineants/api/domain/portfolio/domain/entity/Portfolio.java
@@ -325,9 +325,7 @@ public class Portfolio extends BaseEntity implements Notifiable {
 		return manager.hasTargetGainSendHistory(id);
 	}
 
-	public boolean hasMaxLossSentHistory(NotificationSentRepository manager) {
-		return manager.hasMaxLossSendHistory(id);
-	}
+	//== Portfolio 계산 메서드 시작 ==//
 
 	/**
 	 * 포트폴리오의 총 손익을 계산 후 반환.
@@ -336,13 +334,16 @@ public class Portfolio extends BaseEntity implements Notifiable {
 	 * @return 포트폴리오 총 손익
 	 * @throws IllegalStateException 포트폴리오 종목(PortfolioHolding)에 따른 현재가가 저장소에 없으면 예외 발생
 	 */
-	//== Portfolio 계산 메서드 시작 ==//
 	public Expression calTotalGain(PortfolioCalculator calculator) {
 		return calculator.calTotalGainBy(Collections.unmodifiableList(portfolioHoldings));
 	}
 
+	public boolean hasMaxLossSentHistory(NotificationSentRepository manager) {
+		return manager.hasMaxLossSendHistory(id);
+	}
+
 	public Expression calTotalInvestment(PortfolioCalculator calculator) {
-		return calculator.calTotalInvestment(Collections.unmodifiableList(portfolioHoldings));
+		return calculator.calTotalInvestmentOfHolding(Collections.unmodifiableList(portfolioHoldings));
 	}
 
 	/**
