+++ src/test/java/co/fineants/api/domain/holding/domain/entity/PortfolioHoldingTest.java
@@ -53,29 +53,6 @@ class PortfolioHoldingTest extends AbstractContainerBaseTest {
 		assertThat(totalInvestmentAmount).isEqualByComparingTo(Money.won(100000L));
 	}
 
-	@DisplayName("한 종목의 평균 매입가를 계산한다")
-	@Test
-	void calculateAverageCostPerShare() {
-		// given
-		Portfolio portfolio = createPortfolio(createMember());
-		Stock stock = createSamsungStock();
-		PortfolioHolding portFolioHolding = PortfolioHolding.of(portfolio, stock);
-
-		PurchaseHistory purchaseHistory1 = createPurchaseHistory(null, LocalDateTime.now(), Count.from(5),
-			Money.won(10000), "첫구매", portFolioHolding);
-		PurchaseHistory purchaseHistory2 = createPurchaseHistory(null, LocalDateTime.now(), Count.from(5),
-			Money.won(10000), "첫구매", portFolioHolding);
-
-		portFolioHolding.addPurchaseHistory(purchaseHistory1);
-		portFolioHolding.addPurchaseHistory(purchaseHistory2);
-
-		// when
-		Expression money = portFolioHolding.calculateAverageCostPerShare();
-
-		// then
-		assertThat(money).isEqualByComparingTo(Money.won(10000.0));
-	}
-
 	@DisplayName("한 종목의 총 손익을 계산한다")
 	@Test
 	void calculateTotalGain() {
